requirements.txt
onnxruntime-gpu
nibabel
dicom2nifti
scipy
numpy
